# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.RoleInput:
    properties:
      description:
        maxLength: 500
        type: string
      is_system_role:
        type: boolean
      level:
        minimum: 1
        type: integer
      name:
        maxLength: 50
        minLength: 2
        type: string
      permission_ids:
        items:
          type: integer
        type: array
    required:
    - level
    - name
    type: object
  dto.RolePermissionInput:
    properties:
      permission_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - permission_ids
    type: object
...
  models.Permission:
    properties:
      action:
        maxLength: 50
        minLength: 2
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      resource:
        maxLength: 50
        minLength: 2
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
    required:
    - action
    - name
    - resource
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_system_role:
        type: boolean
      level:
        minimum: 1
        type: integer
      name:
        maxLength: 50
        minLength: 2
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - level
    - name
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        maxLength: 255
        type: string
      full_name:
        maxLength: 255
        minLength: 2
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      uid:
        type: string
      updated_at:
        type: string
    required:
    - email
    - full_name
    - role_id
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@airline.com
    name: API Support
  description: REST API untuk sistem Passenger Service System (PSS) maskapai penerbangan.
    Mencakup manajemen penerbangan, pemesanan tiket, check-in, bagasi, dan boarding
    pass.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Passenger Service Backend API
  version: "1.0"
paths:
...
  /roles:
    get:
      description: Mengambil semua role beserta permission masing-masing. Hanya admin
        ke atas.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Role'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Daftar semua role
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Membuat role baru beserta permission. Hanya super_admin.
      parameters:
      - description: Data role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RoleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Buat role baru
      tags:
      - Role
  /roles/{id}:
    delete:
      description: Menghapus role. System role dan role yang masih digunakan user
        tidak bisa dihapus. Hanya super_admin.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Hapus role
      tags:
      - Role
    get:
      description: Mengambil detail role beserta seluruh permission-nya. Hanya admin
        ke atas.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Detail role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Memperbarui informasi role. Hanya super_admin.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Data update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RoleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Role
  /roles/{id}/permissions:
    post:
      consumes:
      - application/json
      description: Menambahkan satu atau lebih permission ke role. Hanya super_admin.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Daftar permission ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RolePermissionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Tambah permission ke role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Mengganti seluruh permission role dengan daftar baru. Hanya super_admin.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Daftar permission ID baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RolePermissionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Ganti permission role
      tags:
      - Role
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"